More rate limiter tests + code review edits
