+++ scalyr_agent/configuration.py
@@ -28,7 +28,8 @@ import scalyr_agent.util as scalyr_util
 
 from scalyr_agent.json_lib import JsonConversionException, JsonMissingFieldException
 from scalyr_agent.json_lib.objects import JsonObject, JsonArray, ArrayOfStrings
-from scalyr_agent.util import JsonReadFileException, BlockingRateLimiter
+from scalyr_agent.monitor_utils.blocking_rate_limiter import BlockingRateLimiter
+from scalyr_agent.util import JsonReadFileException
 from scalyr_agent.config_util import BadConfiguration, get_config_from_env
 
 from __scalyr__ import get_install_root
