+++ scalyr_agent/monitor_utils/k8s.py
@@ -1503,6 +1503,71 @@ class KubernetesApi( object ):
             finally:
                 rate_limiter.release_token(token, rate_limit_outcome)
 
+    def __open_api_response_log(self, rate_limited):
+        """Opens a file for logging the api response
+
+        The file will be located in agent_log_dir/kapi/(limited/not_limited) depending on whether the
+        api call is rate limited or not.
+
+        @param rate_limited: Whether the response is rate limited or not
+        @type rate_limited: bool
+
+        @returns File handle to the api response log file or None upon failure.
+        @rtype: file handle
+        """
+        # try to open the logged_response_file
+        try:
+            kapi = os.path.join(self.agent_log_path, 'kapi')
+            if not os.path.exists(kapi):
+                os.mkdir(kapi, 0755)
+            if rate_limited:
+                kapi = os.path.join(kapi, 'limited')
+            else:
+                kapi = os.path.join(kapi, 'limited')
+            if not os.path.exists(kapi):
+                os.mkdir(kapi, 0755)
+            fname = '%s_%.20f_%s_%s' % (
+            strftime('%Y%m%d:%H:%M:%S', gmtime()), time.time(), random.randint(1, 100), path.replace('/', '--'))
+
+            # if logging responses to disk, always prepend the stack trace for easier debugging
+            return open(os.path.join(kapi, fname), 'w')
+        except IOError:
+            pass
+
+    def __check_for_fake_response(self, logged_response_file):
+        """Helper method that checks for a well known file on disk and simulates timeouts from API master
+
+        If successfully logging responses, we can also check for a local "simfile" to simulate API master timeout.
+        The simfile is a textfile that contains the HTTP error code we want to simulate.
+
+        This method should only be called during development.
+
+        @param logged_response_file: Logged-response logfile
+        @type logged_response_file: file handle
+
+        @raises K8sApiTemporaryError: if the simfile contains one of the following http error codes that we consider
+            as a temporary error.
+        """
+        fake_response_file = os.path.join(self.agent_log_path, 'simfile')
+        if os.path.isfile(fake_response_file):
+            fake_response_code = None
+            try:
+                fake_f = open(fake_response_file, 'r')
+                try:
+                    fake_response_code = fake_f.read().strip()
+                finally:
+                    fake_f.close()
+            except Exception:
+                if logged_response_file:
+                    logged_response_file.write(
+                        'Error encountered while attempting to fake a response code:\n%s\n\n'
+                        % traceback.format_exc())
+            if fake_response_code in ['404', '503', '429']:
+                global_log.log(scalyr_logging.DEBUG_LEVEL_3,
+                               "Faking api master temporary error (%s) for url: %s",
+                               limit_once_per_x_secs=300, limit_key='k8s_api_query_fake_temporary_error')
+                raise K8sApiTemporaryError('Fake %s' % fake_response_code)
+
     def query_api( self, path, pretty=0, return_temp_errors=False, rate_limited=False):
         """ Queries the k8s API at 'path', and converts OK responses to JSON objects
         """
@@ -1521,53 +1586,16 @@ class KubernetesApi( object ):
         log_responses = self.log_api_responses_to_disk and self.agent_log_path
         logged_response_file = None
         if log_responses:
-            # try to open the logged_response_file
-            try:
-                kapi = os.path.join(self.agent_log_path, 'kapi')
-                if not os.path.exists(kapi):
-                    os.mkdir(kapi, 0755)
-                if rate_limited:
-                    kapi = os.path.join(kapi, 'limited')
-                else:
-                    kapi = os.path.join(kapi, 'limited')
-                if not os.path.exists(kapi):
-                    os.mkdir(kapi, 0755)
-                fname = '%s_%.20f_%s_%s' % (strftime('%Y%m%d:%H:%M:%S', gmtime()), time.time(), random.randint(1, 100), path.replace('/', '--'))
-
-                # if logging responses to disk, always prepend the stack trace for easier debugging
-                logged_response_file = open(os.path.join(kapi, fname), 'w')
-            except IOError:
-                pass
-
+            logged_response_file = self.__open_api_response_log(rate_limited)
         try:
             # Optionally prepend stack trace into logged response file
             if logged_response_file:
                 traceback.print_stack(file=logged_response_file)
                 logged_response_file.write('\n\n')
 
-            # echee: TODO: remove this before shipping
-            # If logging responses, also check for a local "simfile" to simulate API master timeout.
-            # The simfile is a textfile that contains the HTTP error code we want to simulate.
-            if log_responses:
-                fake_response_file = os.path.join(self.agent_log_path, 'simfile')
-                if os.path.isfile(fake_response_file):
-                    fake_response_code = None
-                    try:
-                        fake_f = open(fake_response_file, 'r')
-                        try:
-                            fake_response_code = fake_f.read().strip()
-                        finally:
-                            fake_f.close()
-                    except Exception:
-                        if logged_response_file:
-                            logged_response_file.write(
-                                'Error encountered while attempting to fake a response code:\n%s\n\n'
-                                % traceback.format_exc())
-                    if fake_response_code in ['404', '503', '429']:
-                        global_log.log(scalyr_logging.DEBUG_LEVEL_3,
-                                       "Faking api master temporary error (%s) for url: %s",
-                                       limit_once_per_x_secs=300, limit_key='k8s_api_query_fake_temporary_error')
-                        raise K8sApiTemporaryError('Fake %s' % fake_response_code)
+            # echee: TODO: remove this before merging
+            if logged_response_file:
+                self.__check_for_fake_response(logged_response_file)
 
             # Make actual API call
             try:
